Improve bottom navigation to be visible across all main screens
